+++ setup_template.py
@@ -32,6 +32,8 @@ setuptools.setup(
         "SQLAlchemy-Utils==0.37.8",
         "SQLAlchemy>=2.0.37",
         "GeoAlchemy2>=0.17.0",
+        "orjson>=3.11.3",
+        "loguru>=0.7.3"
     ],
     packages=setuptools.find_packages(where="src"),
     python_requires=">=3.6",