+++ src/pumpwood_flaskviews/views.py
@@ -28,6 +28,8 @@ from pumpwood_flaskviews.query import SqlalchemyQueryMisc
 from pumpwood_flaskviews.auth import AuthFactory
 from pumpwood_flaskviews.action import load_action_parameters
 from pumpwood_i8n.singletons import pumpwood_i8n as _
+from pumpwood_database_error import (
+    TreatPsycopg2Error, TreatSQLAlchemyError)
 
 
 class PumpWoodFlaskView(View):
@@ -395,7 +397,20 @@ class PumpWoodFlaskView(View):
 
         # Save end-points
         if end_point == 'save' and request.method.lower() in ('post', 'put'):
-            return jsonify(self.save(data=data))
+            fields = json.loads(
+                request.args.get('fields', 'null'))
+            foreign_key_fields = json.loads(
+                request.args.get('foreign_key_fields', 'false'))
+            related_fields = json.loads(
+                request.args.get('related_fields', 'false'))
+            default_fields = json.loads(
+                request.args.get('default_fields', 'false'))
+            save_data = self.save(
+                data=data, fields=fields,
+                foreign_key_fields=foreign_key_fields,
+                related_fields=related_fields,
+                default_fields=default_fields)
+            return jsonify(save_data)
 
         if end_point == "save-file-streaming" and \
                 request.method.lower() in ('post', 'put'):
@@ -404,13 +419,24 @@ class PumpWoodFlaskView(View):
                     "Save file stream endpoint have a pk")
 
             # Get URL parameters for the end-point
+            fields = json.loads(
+                request.args.get('fields', 'null'))
+            foreign_key_fields = json.loads(
+                request.args.get('foreign_key_fields', 'false'))
+            related_fields = json.loads(
+                request.args.get('related_fields', 'false'))
+            default_fields = json.loads(
+                request.args.get('default_fields', 'false'))
             file_field = request.args.get('file_field')
             if file_field is None:
                 raise exceptions.PumpWoodForbidden(
                     "file_field not set as url parameter")
             file_name = request.args.get('file_name')
-            return jsonify(self.save_file_streaming(
-                pk=first_arg, file_field=file_field, file_name=file_name))
+            save_streaming_data = self.save_file_streaming(
+                pk=first_arg, file_field=file_field, file_name=file_name,
+                fields=fields, foreign_key_fields=foreign_key_fields,
+                related_fields=related_fields, default_fields=default_fields)
+            return jsonify(save_streaming_data)
 
         if end_point == "remove-file-field" and \
                 request.method.lower() in ('delete'):
@@ -872,15 +898,9 @@ class PumpWoodFlaskView(View):
             try:
                 session.delete(model_object)
                 session.commit()
-            except sqlalchemy.exc.IntegrityError as e:
-                session.rollback()
-                raise exceptions.PumpWoodIntegrityError(message=str(e))
-            except psycopg2.errors.IntegrityError as e:
-                session.rollback()
-                raise exceptions.PumpWoodIntegrityError(message=str(e))
             except Exception as e:
                 session.rollback()
-                raise exceptions.PumpWoodObjectDeleteException(message=str(e))
+                raise e
 
         available_microservices = self.get_available_microservices()
         pumpwood_etl_ok = 'pumpwood-etl-app' in available_microservices
@@ -919,12 +939,38 @@ class PumpWoodFlaskView(View):
             query_result.delete(synchronize_session='fetch')
             session.commit()
         except Exception as e:
-            raise exceptions.PumpWoodObjectDeleteException(message=str(e))
+            session.rollback()
+            raise e
         return True
 
-    def save(self, data, file_paths: dict = {}):
-        """Update object or save new object."""
-        retrieve_serializer = self.serializer(many=False)
+    def save(self, data: dict, file_paths: dict = {},
+             foreign_key_fields: bool = False, related_fields: bool = False,
+             default_fields: bool = False, fields: list = None) -> dict:
+        """Update object or save new object.
+
+        Args:
+            data (dict):
+                Data used to save information on Pumpwood.
+            file_paths (dict):
+                Used when saving files with streaming, it is not
+                exposed to save API. If will set the file path directly on
+                the object.
+            fields (list):
+                Set fields to be returned at serializer.
+            foreign_key_fields (bool):
+                If foreign fields should be returned at the object serializer.
+            related_fields (bool):
+                If related fields should be returned at the object serializer.
+            default_fields (bool):
+                If default fields should be returned at the object serializer.
+
+        Returns:
+            Returns a dictonary with serialized object.
+        """
+        retrieve_serializer = self.serializer(
+            many=False, fields=fields, default_fields=default_fields,
+            foreign_key_fields=foreign_key_fields,
+            related_fields=related_fields)
         retrieve_serializer.context['authorization_token'] = \
             request.headers.get('Authorization', None)
 
@@ -954,15 +1000,9 @@ class PumpWoodFlaskView(View):
             # persist on database.
             session.add(to_save_obj)
             session.flush()
-        except sqlalchemy.exc.IntegrityError as e:
-            session.rollback()
-            raise exceptions.PumpWoodIntegrityError(message=str(e))
-        except psycopg2.errors.IntegrityError as e:
-            session.rollback()
-            raise exceptions.PumpWoodIntegrityError(message=str(e))
         except Exception as e:
             session.rollback()
-            raise exceptions.PumpWoodException(message=str(e))
+            raise e
 
         # Set file names with file_paths dict which is not exposed to API
         # this is only used by save_file_streaming to set file name
@@ -1069,15 +1109,9 @@ class PumpWoodFlaskView(View):
         # information if present.
         try:
             session.commit()
-        except sqlalchemy.exc.IntegrityError as e:
-            session.rollback()
-            raise exceptions.PumpWoodIntegrityError(message=str(e))
-        except psycopg2.errors.IntegrityError as e:
-            session.rollback()
-            raise exceptions.PumpWoodIntegrityError(message=str(e))
         except Exception as e:
             session.rollback()
-            raise exceptions.PumpWoodException(message=str(e))
+            raise e
 
         # Serialize object to return
         result = retrieve_serializer.dump(to_save_obj)
@@ -1106,7 +1140,11 @@ class PumpWoodFlaskView(View):
         return result
 
     def save_file_streaming(self, pk: int | str, file_field: str,
-                            file_name: str = None):
+                            file_name: str = None,
+                            foreign_key_fields: bool = False,
+                            related_fields: bool = False,
+                            default_fields: bool = False,
+                            fields: list = None) -> dict:
         """Save file to object.
 
         Args:
@@ -1116,6 +1154,17 @@ class PumpWoodFlaskView(View):
                 Name of the file field in the object.
             file_name (str):
                 File name that will be set for file streaming.
+            fields (list):
+                Set fields to be returned at serializer.
+            foreign_key_fields (bool):
+                If foreign fields should be returned at the object serializer.
+            related_fields (bool):
+                If related fields should be returned at the object serializer.
+            default_fields (bool):
+                If default fields should be returned at the object serializer.
+
+        Returns:
+            Serialized object.
         """
         if file_field not in self.file_fields.keys():
             raise exceptions.PumpWoodForbidden(
@@ -1808,15 +1857,9 @@ class PumpWoodDataFlaskView(PumpWoodFlaskView):
             try:
                 session.bulk_save_objects(objects_to_load)
                 session.commit()
-            except sqlalchemy.exc.IntegrityError as e:
-                session.rollback()
-                raise exceptions.PumpWoodIntegrityError(message=str(e))
-            except psycopg2.errors.IntegrityError as e:
-                session.rollback()
-                raise exceptions.PumpWoodIntegrityError(message=str(e))
             except Exception as e:
                 session.rollback()
-                raise exceptions.PumpWoodException(message=str(e))
+                raise e
 
             return {'saved_count': len(objects_to_load)}
         else:
@@ -1989,66 +2032,31 @@ def register_pumpwood_view(app, view, service_object: dict):
         return response
 
     # SQLAlchemy errors
-    @app.errorhandler(sqlalchemy.exc.ProgrammingError)
+    @app.errorhandler(sqlalchemy.exc.SQLAlchemyError)
     def handle_sqlalchemy_programmingerror_errors(error):
-        pump_exc = exceptions.PumpWoodException(message=str(error))
-        response = jsonify(pump_exc.to_dict())
-        response.status_code = pump_exc.status_code
-        return response
+        error_dict = TreatSQLAlchemyError.treat(
+            error=error, connection_url=app.config['SQLALCHEMY_DATABASE_URI'])
+        ErrorClass = exceptions.exceptions_dict.get(error_dict['type']) # NOQA
+        if ErrorClass is None:
+            msg = (
+                "Error class returned by 'TreatSQLAlchemyError' [{type}] "
+                "is not implemented on PumpwoodCommunication package.")\
+                .format(type=error_dict['type'])
+            raise NotImplementedError(msg)
 
-    @app.errorhandler(sqlalchemy.exc.IntegrityError)
-    def handle_sqlalchemy_invalidrequest_error(error):
-        pump_exc = exceptions.PumpWoodDatabaseError(message=str(error))
-        response = jsonify(pump_exc.to_dict())
-        response.status_code = pump_exc.status_code
-        return response
-
-    @app.errorhandler(sqlalchemy.exc.InvalidRequestError)
-    def handle_sqlalchemy_invalidrequesterror_error(error):
-        pump_exc = exceptions.PumpWoodQueryException(message=str(error))
+        pump_exc = ErrorClass(
+            message=error_dict['message'], payload=error_dict['payload'])
         response = jsonify(pump_exc.to_dict())
         response.status_code = pump_exc.status_code
         return response
 
     # psycopg2 error handlers
-    @app.errorhandler(psycopg2.errors.DatabaseError)
-    def handle_psycopg2_databaseerror(error):
-        pump_exc = exceptions.PumpWoodDatabaseError(message=str(error))
-        response = jsonify(pump_exc.to_dict())
-        response.status_code = pump_exc.status_code
-        return response
-
-    @app.errorhandler(psycopg2.errors.OperationalError)
-    def handle_psycopg2_operationalerror(error):
-        pump_exc = exceptions.PumpWoodDatabaseError(message=str(error))
-        response = jsonify(pump_exc.to_dict())
-        response.status_code = pump_exc.status_code
-        return response
-
-    @app.errorhandler(psycopg2.errors.NotSupportedError)
-    def handle_psycopg2_notsupportederror(error):
-        pump_exc = exceptions.PumpWoodDatabaseError(message=str(error))
-        response = jsonify(pump_exc.to_dict())
-        response.status_code = pump_exc.status_code
-        return response
-
-    @app.errorhandler(psycopg2.errors.ProgrammingError)
-    def handle_psycopg2_programmingerror(error):
-        pump_exc = exceptions.PumpWoodDatabaseError(message=str(error))
-        response = jsonify(pump_exc.to_dict())
-        response.status_code = pump_exc.status_code
-        return response
-
-    @app.errorhandler(psycopg2.errors.DataError)
-    def handle_psycopg2_dataerror(error):
-        pump_exc = exceptions.PumpWoodDatabaseError(message=str(error))
-        response = jsonify(pump_exc.to_dict())
-        response.status_code = pump_exc.status_code
-        return response
-
-    @app.errorhandler(psycopg2.errors.IntegrityError)
-    def handle_psycopg2_integrityerror(error):
-        pump_exc = exceptions.PumpWoodDatabaseError(message=str(error))
+    @app.errorhandler(psycopg2.Error)
+    def handle_psycopg2_error(error):
+        error_dict = TreatPsycopg2Error.treat(error=error)
+        ErrorClass = exceptions.exceptions_dict.get(error_dict['type']) # NOQA
+        pump_exc = ErrorClass(
+            message=error_dict['message'], payload=error_dict['payload'])
         response = jsonify(pump_exc.to_dict())
         response.status_code = pump_exc.status_code
         return response